Don’t link against -lcrypt or -lcrypto. Preliminary Makefile cleanup.
Since all of the cryptographic primitives required by asleap (DES,
MD4, and SHA1) are obsolete, it is best to ship implementations
ourselves rather than rely on external libraries that may deprecate
them or remove them altogether.  SHA1 is already being supplied by
sha1.c, but MD4 is optionally supplied by OpenSSL (libcrypto) and for
DES we are relying on the obsolete SVID functions ‘encrypt’ and ‘setkey’.

Remove the ifdefs in common.c and genkeys.c allowing for the use of
OpenSSL’s implementation of MD4.  In the Makefile, remove both
-lxcrypt and -lcrypto from LDLIBS, and add md4.o to the link for both
asleap and genkeys.  This means the program will not compile right
now, with undefined references to ‘encrypt’ and ‘setkey’—this will be
corrected in the next commit.

While we’re messing with the Makefile, overhaul it with proper use of
object files, built-in rules and variables, and comprehensive .c:.h
dependency lists.

While we’re messing with the Makefile, eliminate the only place where
we need a -D_LINUX by instead using __GNU_LIBRARY__ to detect the
availability of <endian.h>.  This *may* mean makefile.cygwin is no
longer necessary.  Will come back to that.