While we’re here, some linting.
- Crank up the warnings level (-Wextra -Wpedantic -Wstrict-prototypes)
- Explicitly request C2011 with extensions.
- Explicitly request 64-bit off_t.
- Fix all the resulting warnings.

The only code change of note is that I deleted struct dot11_mgmt from
ieee80211.h.  This was using the old GNU zero-length array extension
in a way that is not supported for C99 flexible array members, and
it’s unused in this program, so the easiest fix is to get rid of it.